-Wextra -pedantic; code cleanup
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@315 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ client/Graphic.cpp
@@ -97,11 +97,11 @@ void Graphic::transform_gl() const {
 	glTranslated(-round(m_center_x), -round(m_center_y), 0.0);
 }
 
-int Graphic::get_image_width() const {
+double Graphic::get_image_width() const {
 	return m_image_width;
 }
 
-int Graphic::get_image_height() const {
+double Graphic::get_image_height() const {
 	return m_image_height;
 }
